updated tests and refactored

+++ app/src/main/java/com/swe6673/finance/resource/FundTransferDOA.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class FundTransferDOA {
 
-	private float amount;
+	private double amount;
 	private String accountNumber;
 	private String routingNumber;
